+++ java-client/src/test/java/org/opensearch/client/documentation/MigrateHlrcTest.java
@@ -36,9 +36,9 @@ import org.opensearch.client.opensearch.OpenSearchClient;
 import org.opensearch.client.json.jackson.JacksonJsonpMapper;
 import org.opensearch.client.transport.OpenSearchTransport;
 import org.opensearch.client.transport.rest_client.RestClientTransport;
-import org.apache.http.HttpHost;
 import org.opensearch.client.RestClient;
 import org.opensearch.client.RestClientBuilder;
+import org.apache.hc.core5.http.HttpHost;
 import org.junit.Test;
 
 public class MigrateHlrcTest {