+++ java-client/src/test/java/org/opensearch/client/opensearch/integTest/PingAndInfoIT.java
@@ -8,7 +8,7 @@
 
 package org.opensearch.client.opensearch.integTest;
 
-import org.apache.http.client.methods.HttpGet;
+import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.opensearch.client.Request;
 import org.opensearch.client.opensearch.OpenSearchClient;
 import org.opensearch.client.opensearch.core.InfoResponse;